+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `getOverview` typescript definition
+- `search` missing typescript definition for resultGroups
 
 ## 0.5.0 - 2022-11-25